feat: Update database, Change Value attribute
- Add Migration to Update database
- Change Value attribute in TransationModel to improve precision

+++ Backend/API/Models/Requests/TransationRequest.cs
@@ -1,7 +1,7 @@
 ﻿namespace Person.Models.Requests
 {
-    public record TransationRequest(string Description , string Status, int Value, Guid PersonId, DateTime? Date);
+    public record TransationRequest(string Description , string Status, decimal Value, Guid PersonId, DateTime? Date);
 
-    public record TransationUpdateRequest(string Description, string Status, int Value, DateTime? Date);
+    public record TransationUpdateRequest(string Description, string Status, decimal Value, DateTime? Date);
 
 }
